fix(flashnode): code check#23131609
