+++ flashnode/cachengine/block.go
@@ -26,7 +26,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/cubefs/cubefs/flashnode"
 	"github.com/cubefs/cubefs/proto"
 	"github.com/cubefs/cubefs/util"
 	"github.com/cubefs/cubefs/util/auditlog"
@@ -579,8 +578,8 @@ func (cb *CacheBlock) InitForCacheRead(sources []*proto.DataSource, readDataNode
 				return e
 			}
 			offset += size
-			flashnode.UpdateWriteBytesMetric(uint64(size), cb.GetRootPath())
-			flashnode.UpdateWriteCountMetric(cb.GetRootPath())
+			updateWriteBytesMetric(uint64(size), cb.GetRootPath())
+			updateWriteCountMetric(cb.GetRootPath())
 			return nil
 		}
 		logPrefix := func() string {
