+++ flashnode/metric.go
@@ -1,11 +1,11 @@
 package flashnode
 
 import (
-	"github.com/cubefs/cubefs/flashnode/cachengine"
 	"path"
 	"sync/atomic"
 	"time"
 
+	"github.com/cubefs/cubefs/flashnode/cachengine"
 	"github.com/cubefs/cubefs/util/exporter"
 	"github.com/cubefs/cubefs/util/log"
 )
@@ -19,15 +19,6 @@ const (
 	MetricFlashNodeHitRate    = "flashNodeHitRate"
 	MetricFlashNodeEvictCount = "flashNodeEvictCount"
 )
-
-var StatMap = make(map[string]*MetricStat)
-
-type MetricStat struct {
-	ReadBytes  uint64
-	ReadCount  uint64
-	WriteBytes uint64
-	WriteCount uint64
-}
 
 type FlashNodeMetrics struct {
 	flashNode        *FlashNode
@@ -53,7 +44,7 @@ func (f *FlashNode) registerMetrics(disks []*cachengine.Disk) {
 	f.metrics.MetricEvictCount = exporter.NewGauge(MetricFlashNodeEvictCount)
 	f.metrics.MetricHitRate = exporter.NewGauge(MetricFlashNodeHitRate)
 	for _, d := range disks {
-		StatMap[path.Join(d.Path, cachengine.DefaultCacheDirName)] = new(MetricStat)
+		cachengine.StatMap[path.Join(d.Path, cachengine.DefaultCacheDirName)] = new(cachengine.MetricStat)
 	}
 
 	log.LogInfof("registerMetrics")
@@ -90,28 +81,28 @@ func (fm *FlashNodeMetrics) doStat() {
 }
 
 func (fm *FlashNodeMetrics) setReadBytesMetric() {
-	for d, stat := range StatMap {
+	for d, stat := range cachengine.StatMap {
 		readBytes := atomic.SwapUint64(&stat.ReadBytes, 0)
 		fm.MetricReadBytes.SetWithLabels(float64(readBytes), map[string]string{"cluster": fm.flashNode.clusterID, exporter.FlashNode: fm.flashNode.localAddr, exporter.Disk: d})
 	}
 }
 
 func (fm *FlashNodeMetrics) setReadCountMetric() {
-	for d, stat := range StatMap {
+	for d, stat := range cachengine.StatMap {
 		readCount := atomic.SwapUint64(&stat.ReadCount, 0)
 		fm.MetricReadCount.SetWithLabels(float64(readCount), map[string]string{"cluster": fm.flashNode.clusterID, exporter.FlashNode: fm.flashNode.localAddr, exporter.Disk: d})
 	}
 }
 
 func (fm *FlashNodeMetrics) setWriteBytesMetric() {
-	for d, stat := range StatMap {
+	for d, stat := range cachengine.StatMap {
 		writeBytes := atomic.SwapUint64(&stat.WriteBytes, 0)
 		fm.MetricWriteBytes.SetWithLabels(float64(writeBytes), map[string]string{"cluster": fm.flashNode.clusterID, exporter.FlashNode: fm.flashNode.localAddr, exporter.Disk: d})
 	}
 }
 
 func (fm *FlashNodeMetrics) setWriteCountMetric() {
-	for d, stat := range StatMap {
+	for d, stat := range cachengine.StatMap {
 		writeCount := atomic.SwapUint64(&stat.WriteCount, 0)
 		fm.MetricWriteCount.SetWithLabels(float64(writeCount), map[string]string{"cluster": fm.flashNode.clusterID, exporter.FlashNode: fm.flashNode.localAddr, exporter.Disk: d})
 	}
@@ -131,26 +122,14 @@ func (fm *FlashNodeMetrics) setHitRateMetric() {
 	}
 }
 
-func UpdateReadBytesMetric(size uint64, d string) {
-	if stat, ok := StatMap[d]; ok {
+func (fm *FlashNodeMetrics) updateReadBytesMetric(size uint64, d string) {
+	if stat, ok := cachengine.StatMap[d]; ok {
 		atomic.AddUint64(&stat.ReadBytes, size)
 	}
 }
 
-func UpdateReadCountMetric(d string) {
-	if stat, ok := StatMap[d]; ok {
+func (fm *FlashNodeMetrics) updateReadCountMetric(d string) {
+	if stat, ok := cachengine.StatMap[d]; ok {
 		atomic.AddUint64(&stat.ReadCount, 1)
-	}
-}
-
-func UpdateWriteBytesMetric(size uint64, d string) {
-	if stat, ok := StatMap[d]; ok {
-		atomic.AddUint64(&stat.WriteBytes, size)
-	}
-}
-
-func UpdateWriteCountMetric(d string) {
-	if stat, ok := StatMap[d]; ok {
-		atomic.AddUint64(&stat.WriteCount, 1)
 	}
 }
